backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "verve-fds-backend"
version = "0.1.0"
description = "VERVE FDS Infra backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.32,<3.0.0",
  "psycopg[binary]>=3.2.1,<4.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "PyJWT>=2.9.0,<3.0.0",
  "redis>=5.0.8,<6.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2,<9.0.0",
  "httpx>=0.27.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

